Stop using __enabled to determine FileOpenHook state.
Previously we used a boolean flag FileOpenHook.__enabled to determine
whether the hook was monitoring an event. This could lead to issues if
nested events used the hook, since as soon as we exited the nested event
the hook would be disabled, even if we were still executing the parent
event.

Instead, we now look at the __current_event_stack to see if it's
non-empty. If it is, we now that FileOpenHook is executing an event, so
we should consider it to be enabled.