Fix the logger keyword argument for Auditor.
Previously, there was a bug where Auditor would not properly respect the
"logger" keyword argument to Auditor.__init__ if the logger was
specified as a string. Instead, Auditor would simply revert to using the
default "seagrass" auditor.

This commit fixes this bug and adds more tests to check that Auditor
respects the logger keyword argument.