Add the CleanupHook interface.
Define a new interface for hooks that should have a 'clean-up' stage run
unconditionally before an event finishes executing. This is useful for
cases where a hook may introduce side effects on internal or global
state that should be rectified before we finish the event.

The two hooks seagrass.hooks.FileOpenHook and
seagrass.hooks.ProfilerHook now both implement this interface.

Closes #35.