Add an additional handler to the test logger for warnings and errors.
Add a handler to the logging configuration set by SeagrassTestCaseMixin
that emits all logs with level >= WARNING to stdout. This way we can see
warnings and errors that are emitted by the logger during failing test
cases, which can provide additional information we can use to debug
them.

The new handler can be disabled by setting the log_warnings_to_stdout
attribute to False in test cases that use SeagrassTestCaseMixin.