Modify FileOpenHook to be a child class of RuntimeAuditHook.
Have FileOpenHook inherit from RuntimeAuditHook, which provides a few
advantages:
- It massively simplifies the implementation of FileOpenHook (we cut the
  entire file down to almost half its original size).
- FileOpenHook now serves as a good template for anybody looking to
  implement a hook that inherits from RuntimeAuditHook.
- We carry some features of RuntimeAuditHook down into FileOpenHook. For
  instance, we can use the improved error handling provided by
  RuntimeAuditHook when an error arises in FileOpenHook.sys_hook.

One drawback of this change is that we no longer keep track of the
Seagrass event stack (which is doable, but it would require adding a bit
of extra code to FileOpenHook that we don't really want or need). As a
result, the track_nested_hooks flag has been removed from FileOpenHook's
constructor.