Fix issue where ProfilerHook would be prematurely disabled.
Fix a bug where ProfilerHook is disabled every time cleanup() gets
called, which causes us to miss profiling information in the outer event
of a pair of nested events. Previously, profiling would be disabled once
we exited the inner event, which means that if there was any code
between the end of the inner event and the end of the outer event, the
profiler would completely miss it.