Make it possible to filter events that get audited by name.
Add an event_filter property to Auditor that filters which events should
be audited and which should not. event_filter is a function taking a
single string corresponding to the event name, and outputs True or False
based on whether that event should be audited.

In addition to setting event_filter manually, users can now pass in the
'filter' keyword argument to start_auditing to create an auditing
context in which only the filtered events are audited.

Closes #42.