Make it possible to auto-name Seagrass events.
Allow Seagrass to accept a function for the event name used by
Auditor.audit, which takes a function and generates an event name from
it. Add an 'auto' function to seagrass that can be used to automatically
generate event names for functions, e.g.

    import seagrass

    @seagrass.audit(seagrass.auto)
    def foo(*args, **kwargs):
        ...

The auto-generated name will be based on the path to the audited
function.

Closes #50.