Allow audit hooks set by RuntimeAuditHook to be traced.
Per the documentation for sys.addaudithook, audit hooks cannot be traced
by sys.settrace unless they have a __cantrace__ attribute that is set to
True. This commit adds a new keyword argument to the RuntimeAuditHook
constructor that allows users to determine whether or not they want the
hook to be traceable.