Add compatibility with Python 3.7.
Make various adjustments so that Seagrass is compatible with Python 3.7.
We make two changes to provide type annotations that are supported in
Python >= 3.8 but not for earlier versions:

- Use typing_extensions to import various type annotations that Seagrass
  uses that were unavailable in Python 3.7. Note that the
  typing_extensions dependency is only installed for Python < 3.8; for
  later versions it is skipped over in the requirements.txt.
- Instead of importing type annotation classes from `typing`, they're
  now imported from the new seagrass._typing. This allows us to
  dynamically import annotations via typing_extensions for Python 3.7.

In addition, I've stopped using the following features which were also
unavailable before Python 3.8:

- Stop using the walrus operator (:=) to make assignments.
- Stop using f-strings patterned like f"{var=}" as shorthand for
  f"var={var}".

Finally, in cases where we have to use sys.audit or sys.addaudithook,
Seagrass will raise a NotImplementedError if sys.audit is unavailable.
The test suite skips the test cases that explicitly depend on the
existence of runtime audit hooking for Python < 3.8, and there are new
tests to ensure that NotImplementError will be raised for these
versions.