Modify the RuntimeAuditHook and TracingHook API.
Change RuntimeAuditHook and TracingHook so that each of them can be
instantiated, accepting a single function that obeys the
sys.addaudithook or sys.settrace API, respectively. This makes it
possible to create new RuntimeAuditHooks and TracingHooks without having
to subclass those types, although it is still fairly straightforward to
create one of these hooks by doing so.