Add type annotations for the exported global auditor attributes.
Add annotations for the methods and properties of the global_auditor()
that get exported from seagrass/__init__.py, e.g. audit(),
create_event(), and so on. In addition, the new method for retrieving
these attributes is significantly streamlined, and no longer relies on
complex caching mechanisms plus the use of the __getattr__ function.