Fix issue where comparison with t.MISSING could raise an exception.
In limited circumstances, making a comparison `x == t.MISSING` or
`x != t.MISSING` could raise an exception, especially when the type of
`x` is completely outside the scope of Seagrass. Namely, if `x`
overrides `__eq__` or `__neq__`, it has to be able to accept arguments
of type `t.Missing` in order for this comparison to work.

I found that in practice, some classes that override these methods
expect their argument to be of the same class, which can trigger
hard-to-diagnose errors. This commit fixes the problem by changing
comparisons like `x == t.MISSING` to `isinstance(x, t.Missing)`, which
is much less likely to result in an error.

Closes #66.