Declare that the package includes type annotations.
Modify the setup.py to declare that the package includes type
annotations by adding a py.typed file and by adding the "package_data"
key to the call to setup(). This should ensure that packages that depend
on Seagrass are able to find type annotations when running mypy.

Bump version to v0.9.1.